Move data file handlers out of main
