+++ src/main.ts
@@ -1,15 +1,12 @@
 import * as Sentry from '@sentry/electron/main'
 import { watch, FSWatcher } from 'chokidar'
-import { COPYFILE_EXCL } from 'constants'
-import { app, BrowserWindow, dialog, ipcMain, nativeTheme } from 'electron'
+import { app, BrowserWindow, ipcMain, nativeTheme } from 'electron'
 import log from 'electron-log/main'
 import { existsSync } from 'fs'
-import { copyFile, readdir, rename, readFile, stat } from 'fs/promises'
+import { readdir, rename } from 'fs/promises'
 import path from 'path'
-import invariant from 'tiny-invariant'
 import { updateElectronApp } from 'update-electron-app'
 
-import { MAX_DATA_FILE_SIZE } from './constants/files'
 import {
   DATA_FILES_PATH,
   GENERATORS_PATH,
@@ -32,9 +29,7 @@ import {
 import { BrowserServer } from './services/browser/server'
 import { getSettings, initSettings, saveSettings } from './settings'
 import { ProxyStatus } from './types'
-import { DataFilePreview } from './types/testData'
 import { sendReport } from './usageReport'
-import { parseDataFile } from './utils/dataFile'
 import {
   getAppIcon,
   getPlatform,
@@ -272,60 +267,6 @@ ipcMain.on('app:close', (event) => {
   const browserWindow = browserWindowFromEvent(event)
   browserWindow.close()
 })
-
-ipcMain.handle('data-file:import', async (event) => {
-  const browserWindow = browserWindowFromEvent(event)
-
-  const dialogResult = await dialog.showOpenDialog(browserWindow, {
-    message: 'Import data file',
-    properties: ['openFile'],
-    filters: [{ name: 'Supported data files', extensions: ['csv', 'json'] }],
-  })
-
-  const filePath = dialogResult.filePaths[0]
-
-  if (dialogResult.canceled || !filePath) {
-    return
-  }
-
-  const { size } = await stat(filePath)
-  invariant(size <= MAX_DATA_FILE_SIZE, 'File is too large')
-
-  await copyFile(
-    filePath,
-    path.join(DATA_FILES_PATH, path.basename(filePath)),
-    COPYFILE_EXCL
-  )
-
-  return path.basename(filePath)
-})
-
-ipcMain.handle(
-  'data-file:load-preview',
-  async (_, fileName: string): Promise<DataFilePreview> => {
-    const fileType = fileName.split('.').pop()
-    const filePath = path.join(DATA_FILES_PATH, fileName)
-
-    invariant(
-      fileType === 'csv' || fileType === 'json',
-      'Unsupported file type'
-    )
-
-    const data = await readFile(filePath, {
-      flag: 'r',
-      encoding: 'utf-8',
-    })
-
-    const parsedData = parseDataFile(data, fileType)
-
-    return {
-      type: fileType,
-      data: parsedData.slice(0, 20),
-      props: parsedData[0] ? Object.keys(parsedData[0]) : [],
-      total: parsedData.length,
-    }
-  }
-)
 
 ipcMain.on('splashscreen:close', (event) => {
   console.info('splashscreen:close event received')
